container/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
httpx==0.28.1
# RAG dependencies
chromadb==1.0.1
anthropic==0.40.0
cairosvg==2.8.2
pillow==11.2.1
